backend/requirements.txt
# Flask Core
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.5

# Database
SQLAlchemy==2.0.23
alembic==1.13.1

# Market Data
yfinance==0.2.33
beautifulsoup4==4.12.2
requests==2.31.0
lxml

# AI Integration
google-generativeai==0.3.2

# Payment
paypalrestsdk==1.13.1

# Background Tasks
APScheduler==3.10.4

# Security
python-dotenv==1.0.0
werkzeug==3.0.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3
